requirements.txt
setuptools==69.5.1
wheel==0.43.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.6.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
numpy==1.24.3
scipy==1.10.1
httpx==0.25.1
gunicorn==21.2.0
chardet==5.2.0
torch==2.0.1
transformers==4.30.2